Revert "init camera in AbstraceCamera"
This reverts commit ef86a44cc976c71ccf3b890a96274f9a2527bbe2.
