+++ system/camerad/cameras/camera_ar0231.cc
@@ -64,11 +64,6 @@ CameraAR0231::~CameraAR0231() {
 
 }
 
-int CameraAR0231::getSlaveAddress(int port) {
-  assert(port >=0 && port <= 2);
-  return (int[]){0x20, 0x30, 0x20}[port];
-}
-
 std::vector<struct i2c_random_wr_payload> CameraAR0231::getExposureVector(int new_g, bool dc_gain_enabled, int exposure_time, int dc_gain_weight) const {
   uint16_t analog_gain_reg = 0xFF00 | (new_g << 4) | new_g;
   return {
@@ -78,7 +73,7 @@ std::vector<struct i2c_random_wr_payload> CameraAR0231::getExposureVector(int ne
   };
 }
 
-std::map<uint16_t, std::pair<int, int>> CameraAR0231::buildREgisterLut(uint8_t *data) {
+std::map<uint16_t, std::pair<int, int>> CameraAR0231::buildRegisterLut(uint8_t *data) {
   // This function builds a lookup table from register address, to a pair of indices in the
   // buffer where to read this address. The buffer contains padding bytes,
   // as well as markers to indicate the type of the next byte.
@@ -132,7 +127,7 @@ std::map<uint16_t, std::pair<int, int>> CameraAR0231::buildREgisterLut(uint8_t *
 
 std::map<uint16_t, uint16_t> CameraAR0231::parseRegisters(uint8_t *data, std::initializer_list<uint16_t> addrs) {
   if (ar0231_register_lut.empty()) {
-    ar0231_register_lut = buildREgisterLut(data);
+    ar0231_register_lut = buildRegisterLut(data);
   }
 
   std::map<uint16_t, uint16_t> registers;
