+++ system/camerad/cameras/camera_qcom2.h
@@ -66,6 +66,7 @@ public:
   bool skipped;
 
   CameraBuf buf;
+  MemoryManager mm;
 
 private:
   void config_isp(int io_mem_handle, int fence, int request_id, int buf0_mem_handle, int buf0_offset);
